[스터디 관리하기 뷰 시작 후] 서버 연결

+++ android/domain/src/main/java/com/created/domain/model/Member.kt
@@ -4,6 +4,6 @@ data class Member(
     val id: Long,
     val tier: Int,
     val nickname: String,
-    val successRate: Int,
+    val successRate: Double,
     val profileImage: String,
 )
